Feature/service 11

+++ Hot4/Hot4.ViewModel/NetworkModel.cs
@@ -2,7 +2,7 @@
 {
     public class NetworkModel
     {
-        public int NetworkId { get; set; }
+        public byte NetworkId { get; set; }
         public string Network { get; set; }
         public string Prefix { get; set; }
     }
